# api_gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway service for payment app",
        "title": "Payment API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/customer/payment/authorize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Authorizes a payment from the authenticated customer to the merchant.\nTransfers the payment amount from the customer's Default account to their Payment account.\nReturns a PID used to capture the payment as the next step.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Authorize a customer payment",
                "parameters": [
                    {
                        "description": "Authorize payment payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.AuthorizePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.AuthorizeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customer/payment/capture": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Captures (finalizes) an authorized payment using its PID.\nSends the payment amount from the customer's Payment account to the merchant's Incoming account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Capture a payment",
                "parameters": [
                    {
                        "description": "Capture payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CapturePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK (empty body)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Login endpoint for a customer. Required before initiating a payment.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Customer login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base64 encoded username and password, e.g., 'Basic base64(username:password)'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.AuthorizePayload": {
            "type": "object",
            "properties": {
                "merchant_user_id": {
                    "type": "string"
                },
                "paise": {
                    "type": "integer"
                }
            }
        },
        "main.AuthorizeResponse": {
            "type": "object",
            "properties": {
                "pid": {
                    "type": "string"
                }
            }
        },
        "main.CapturePayload": {
            "type": "object",
            "properties": {
                "pid": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer \u003cjwt\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}